requirements.txt
pycryptodome
pygame-ce